refactor routine: helper method 분리
