+++ src/main/java/com/groomthon/habiglow/domain/routine/facade/RoutineManagementFacade.java
@@ -1,27 +1,19 @@
 package com.groomthon.habiglow.domain.routine.facade;
 
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.groomthon.habiglow.domain.member.entity.MemberEntity;
-import com.groomthon.habiglow.domain.member.repository.MemberRepository;
 import com.groomthon.habiglow.domain.routine.dto.request.CreateRoutineRequest;
 import com.groomthon.habiglow.domain.routine.dto.request.UpdateRoutineRequest;
 import com.groomthon.habiglow.domain.routine.dto.response.RoutineResponse;
-import com.groomthon.habiglow.domain.routine.entity.RoutineEntity;
-import com.groomthon.habiglow.global.exception.BaseException;
-import com.groomthon.habiglow.global.response.ErrorCode;
-import com.groomthon.habiglow.domain.routine.repository.RoutineRepository;
-import com.groomthon.habiglow.domain.routine.service.RoutineValidationService;
-import com.groomthon.habiglow.domain.routine.service.GrowthAnalysisService;
-import com.groomthon.habiglow.domain.routine.service.ReductionAnalysisService;
-import com.groomthon.habiglow.domain.routine.service.GrowthConfigurationService;
-import com.groomthon.habiglow.domain.routine.dto.response.adaptation.RoutineAdaptationResultResponse;
 import com.groomthon.habiglow.domain.routine.dto.response.adaptation.AdaptationAction;
-import com.groomthon.habiglow.domain.routine.event.RoutineTargetChangedEvent;
-import com.groomthon.habiglow.domain.routine.event.RoutineCreatedEvent;
-import com.groomthon.habiglow.domain.routine.event.RoutineDeletedEvent;
+import com.groomthon.habiglow.domain.routine.dto.response.adaptation.RoutineAdaptationResultResponse;
+import com.groomthon.habiglow.domain.routine.entity.RoutineEntity;
+import com.groomthon.habiglow.domain.routine.repository.RoutineRepository;
+import com.groomthon.habiglow.domain.routine.service.GrowthAnalysisService;
+import com.groomthon.habiglow.domain.routine.service.GrowthConfigurationService;
+import com.groomthon.habiglow.domain.routine.service.ReductionAnalysisService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -32,26 +24,24 @@ import lombok.extern.slf4j.Slf4j;
 public class RoutineManagementFacade {
     
     private final RoutineRepository routineRepository;
-    private final MemberRepository memberRepository;
-    private final RoutineValidationService validationService;
+    private final RoutineManagementHelper helper;
     private final GrowthAnalysisService growthAnalysisService;
     private final ReductionAnalysisService reductionAnalysisService;
     private final GrowthConfigurationService growthConfigurationService;
-    private final ApplicationEventPublisher eventPublisher;
 
     // ==================== LIFECYCLE MANAGEMENT ====================
 
     @Transactional
     public RoutineResponse createRoutineWithFullValidation(Long memberId, CreateRoutineRequest request) {
-        MemberEntity member = findMemberById(memberId);
+        MemberEntity member = helper.findMemberById(memberId);
         
-        validateAndCreateRoutine(request);
+        helper.validateAndCreateRoutine(request);
         
-        RoutineEntity savedRoutine = createAndSaveRoutine(member, request);
+        RoutineEntity savedRoutine = helper.createAndSaveRoutine(member, request);
         
-        handleGrowthModeLogging(savedRoutine, memberId);
+        helper.handleGrowthModeLogging(savedRoutine, memberId);
         
-        publishRoutineCreatedEvent(savedRoutine, memberId);
+        helper.publishRoutineCreatedEvent(savedRoutine, memberId);
         
         log.info("Created routine: {} for member: {}", savedRoutine.getRoutineId(), memberId);
         return RoutineResponse.from(savedRoutine);
@@ -59,9 +49,9 @@ public class RoutineManagementFacade {
 
     @Transactional
     public RoutineResponse updateRoutineWithValidation(Long memberId, Long routineId, UpdateRoutineRequest request) {
-        RoutineEntity routine = findRoutineByIdAndMemberId(routineId, memberId);
+        RoutineEntity routine = helper.findRoutineByIdAndMemberId(routineId, memberId);
         
-        validateAndUpdateRoutine(request);
+        helper.validateAndUpdateRoutine(request);
         
         boolean wasGrowthEnabled = routine.isGrowthModeEnabled();
         
@@ -75,7 +65,7 @@ public class RoutineManagementFacade {
                 request.getTargetIncrement()
         );
         
-        handleGrowthModeChange(routine, wasGrowthEnabled, memberId);
+        helper.handleGrowthModeChange(routine, wasGrowthEnabled, memberId);
         
         log.info("Updated routine: {} for member: {}", routineId, memberId);
         return RoutineResponse.from(routine);
@@ -83,19 +73,13 @@ public class RoutineManagementFacade {
 
     @Transactional
     public void deleteRoutine(Long memberId, Long routineId) {
-        RoutineEntity routine = findRoutineByIdAndMemberId(routineId, memberId);
+        RoutineEntity routine = helper.findRoutineByIdAndMemberId(routineId, memberId);
         
-        // 루틴 삭제 이벤트 발행 (삭제 전에 데이터 수집)
-        RoutineDeletedEvent event = RoutineDeletedEvent.of(
-            routine.getRoutineId(), routine.getTitle(), memberId,
-            routine.getCategory(), routine.isGrowthModeEnabled(), 
-            routine.getTargetValue()
-        );
-        eventPublisher.publishEvent(event);
+        helper.publishRoutineDeletedEvent(routine, memberId);
         
-        handleDeletionLogging(routine, memberId);
+        helper.handleDeletionLogging(routine, memberId);
         
-        routineRepository.delete(routine);
+        helper.deleteRoutine(routine);
     }
 
     // ==================== OPTIMIZATION MANAGEMENT ====================
@@ -105,7 +89,7 @@ public class RoutineManagementFacade {
      */
     @Transactional
     public RoutineAdaptationResultResponse executeRoutineAdaptation(Long memberId, Long routineId, AdaptationAction action) {
-        RoutineEntity routine = findRoutineByIdAndMemberId(routineId, memberId);
+        RoutineEntity routine = helper.findRoutineByIdAndMemberId(routineId, memberId);
         
         return switch (action) {
             case INCREASE -> executeGrowthAdaptation(routine, memberId);
@@ -120,7 +104,7 @@ public class RoutineManagementFacade {
         Integer previousTarget = routine.getTargetValue();
         Integer newTarget = growthConfigurationService.executeTargetIncrease(routine);
         
-        publishTargetChangedEvent(routine, memberId, previousTarget, newTarget, AdaptationAction.INCREASE);
+        helper.publishTargetChangedEvent(routine, memberId, previousTarget, newTarget, AdaptationAction.INCREASE);
         
         log.info("Growth adaptation executed for routine: {} from {} to {} by member: {}",
             routine.getRoutineId(), previousTarget, newTarget, memberId);
@@ -136,7 +120,7 @@ public class RoutineManagementFacade {
         Integer previousTarget = routine.getTargetValue();
         Integer newTarget = growthConfigurationService.executeTargetDecrease(routine);
         
-        publishTargetChangedEvent(routine, memberId, previousTarget, newTarget, AdaptationAction.DECREASE);
+        helper.publishTargetChangedEvent(routine, memberId, previousTarget, newTarget, AdaptationAction.DECREASE);
         
         log.info("Reduction adaptation executed for routine: {} from {} to {} by member: {}",
             routine.getRoutineId(), previousTarget, newTarget, memberId);
@@ -156,117 +140,6 @@ public class RoutineManagementFacade {
         return RoutineAdaptationResultResponse.success(
             routine.getRoutineId(), routine.getTitle(), 
             previousCycleDays, 0, AdaptationAction.RESET);
-    }
-
-    // ==================== PRIVATE HELPER METHODS ====================
-
-    private MemberEntity findMemberById(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new BaseException(ErrorCode.MEMBER_NOT_FOUND));
-    }
-
-    private RoutineEntity findRoutineByIdAndMemberId(Long routineId, Long memberId) {
-        return routineRepository.findByRoutineIdAndMember_Id(routineId, memberId)
-                .orElseThrow(() -> new BaseException(ErrorCode.ROUTINE_NOT_FOUND));
-    }
-
-    private void publishRoutineCreatedEvent(RoutineEntity routine, Long memberId) {
-        RoutineCreatedEvent event = RoutineCreatedEvent.of(
-            routine.getRoutineId(), routine.getTitle(), memberId,
-            routine.getCategory(), routine.getIsGrowthMode(),
-            routine.getTargetType(), routine.getTargetValue()
-        );
-        eventPublisher.publishEvent(event);
-    }
-
-    private void publishRoutineDeletedEvent(RoutineEntity routine, Long memberId) {
-        RoutineDeletedEvent event = RoutineDeletedEvent.of(
-            routine.getRoutineId(), routine.getTitle(), memberId,
-            routine.getCategory(), routine.isGrowthModeEnabled(), 
-            routine.getTargetValue()
-        );
-        eventPublisher.publishEvent(event);
-    }
-
-    private void publishTargetChangedEvent(RoutineEntity routine, Long memberId, 
-                                         Integer previousTarget, Integer newTarget, AdaptationAction action) {
-        RoutineTargetChangedEvent event = RoutineTargetChangedEvent.of(
-            routine.getRoutineId(), routine.getTitle(), memberId,
-            previousTarget, newTarget, action
-        );
-        eventPublisher.publishEvent(event);
-    }
-
-    private RoutineEntity createAndSaveRoutine(MemberEntity member, CreateRoutineRequest request) {
-        RoutineEntity routine = RoutineEntity.createRoutine(
-                member,
-                request.getTitle(),
-                request.getDescription(),
-                request.getCategory(),
-                request.getIsGrowthMode(),
-                request.getTargetType(),
-                request.getTargetValue(),
-                request.getGrowthCycleDays(),
-                request.getTargetIncrement()
-        );
-        
-        return routineRepository.save(routine);
-    }
-
-    private void handleGrowthModeLogging(RoutineEntity routine, Long memberId) {
-        if (routine.isGrowthModeEnabled()) {
-            logGrowthModeActivated(routine, memberId);
-        }
-    }
-
-    private void handleDeletionLogging(RoutineEntity routine, Long memberId) {
-        if (routine.isGrowthModeEnabled()) {
-            log.info("Deleting growth-enabled routine: {} for member: {} - Lost target: {}", 
-                    routine.getRoutineId(), memberId, routine.getTargetValue());
-        } else {
-            log.info("Deleting routine: {} for member: {}", routine.getRoutineId(), memberId);
-        }
-    }
-
-    private void validateAndCreateRoutine(CreateRoutineRequest request) {
-        validationService.validateCreateRequest(
-                request.getTitle(),
-                request.getCategory(),
-                request.getIsGrowthMode(),
-                request.getTargetType(),
-                request.getTargetValue(),
-                request.getGrowthCycleDays(),
-                request.getTargetIncrement()
-        );
-    }
-
-    private void validateAndUpdateRoutine(UpdateRoutineRequest request) {
-        validationService.validateUpdateRequest(
-                request.getDescription(),
-                request.getCategory(),
-                request.getIsGrowthMode(),
-                request.getTargetType(),
-                request.getTargetValue(),
-                request.getGrowthCycleDays(),
-                request.getTargetIncrement()
-        );
-    }
-
-    private void handleGrowthModeChange(RoutineEntity routine, boolean wasGrowthEnabled, Long memberId) {
-        boolean isNowGrowthEnabled = routine.isGrowthModeEnabled();
-        
-        if (!wasGrowthEnabled && isNowGrowthEnabled) {
-            log.info("Growth mode activated for routine: {} by member: {}", 
-                    routine.getRoutineId(), memberId);
-        } else if (wasGrowthEnabled && !isNowGrowthEnabled) {
-            log.info("Growth mode deactivated for routine: {} by member: {}", 
-                    routine.getRoutineId(), memberId);
-        }
-    }
-
-    private void logGrowthModeActivated(RoutineEntity routine, Long memberId) {
-        log.info("New routine with growth mode created: {} for member: {} - Target: {}, Increment: {}", 
-                routine.getRoutineId(), memberId, routine.getTargetValue(), routine.getTargetIncrement());
     }
 
 }
